Move license checker to services in v7/v8
