+++ appsec-kit-v7/src/main/java/com/vaadin/appsec/v7/service/LicenseCheckerWebListener.java
@@ -7,22 +7,26 @@
  * See <https://vaadin.com/commercial-license-and-service-terms> for the full
  * license.
  */
-package com.vaadin.appsec.v8;
+package com.vaadin.appsec.v7.service;
 
 import java.io.IOException;
+import java.util.List;
 import java.util.Properties;
+import java.util.concurrent.CopyOnWriteArrayList;
+
+import javax.servlet.annotation.WebListener;
+import javax.servlet.http.HttpSessionEvent;
+import javax.servlet.http.HttpSessionListener;
 
 import com.vaadin.pro.licensechecker.BuildType;
 import com.vaadin.pro.licensechecker.LicenseChecker;
-import com.vaadin.server.ServiceInitEvent;
 import com.vaadin.server.VaadinService;
-import com.vaadin.server.VaadinServiceInitListener;
 
 /**
- * Service initialization listener to verify the license.
+ * Session listener to verify the license.
  */
-public class LicenseCheckerServiceInitListener
-        implements VaadinServiceInitListener {
+@WebListener
+public class LicenseCheckerWebListener implements HttpSessionListener {
 
     static final String PROPERTIES_RESOURCE = "appsec-kit.properties";
 
@@ -30,14 +34,30 @@ public class LicenseCheckerServiceInitListener
 
     static final String PRODUCT_NAME = "vaadin-appsec-kit";
 
-    @Override
-    public void serviceInit(ServiceInitEvent event) {
-        final VaadinService service = event.getSource();
+    private static final List<String> initializedVaadinServiceNames = new CopyOnWriteArrayList<>();
 
+    @Override
+    public synchronized void sessionCreated(HttpSessionEvent se) {
+        VaadinService vaadinService = VaadinService.getCurrent();
+        if (vaadinService != null) {
+            String serviceName = vaadinService.getServiceName();
+            if (!initializedVaadinServiceNames.contains(serviceName)) {
+                checkLicense(vaadinService);
+                initializedVaadinServiceNames.add(serviceName);
+            }
+        }
+    }
+
+    @Override
+    public void sessionDestroyed(HttpSessionEvent se) {
+        // NO-OP
+    }
+
+    private void checkLicense(VaadinService service) {
         try {
             final Properties properties = new Properties();
-            properties.load(LicenseCheckerServiceInitListener.class
-                    .getClassLoader().getResourceAsStream(PROPERTIES_RESOURCE));
+            properties.load(LicenseCheckerWebListener.class.getClassLoader()
+                    .getResourceAsStream(PROPERTIES_RESOURCE));
             final String version = properties.getProperty(VERSION_PROPERTY);
 
             // Check the license at runtime if in development mode
