+++ appsec-kit-v8/src/main/resources/META-INF/services/com.vaadin.server.VaadinServiceInitListener
@@ -1,3 +1,3 @@
-com.vaadin.appsec.v8.LicenseCheckerServiceInitListener
+com.vaadin.appsec.v8.service.LicenseCheckerServiceInitListener
 com.vaadin.appsec.v8.service.AppSecServiceInitListener
 com.vaadin.appsec.v8.service.NotificationInitListener
